+++ src/lsp/cobol_lsp/lsp_data_info_printer.ml
@@ -182,19 +182,32 @@ and pp_field_layout: field_layout Pretty.printer = fun ppf x ->
       Fmt.const pp_struct subfields ppf x
 
 and pp_field_definition: field_definition Pretty.printer = fun ppf x ->
-  let pp_qualname_opt_in_block' = pp_cobol_block Fmt.(option ~none:(any "FILLER") Cobol_ptree.pp_qualname')  in
-  Fmt.(
-    const pp_qualname_opt_in_block' x.field_qualname
-    ++ any "\n\n"
-    ++ const pp_field_layout x.field_layout
-    ++ (match x.field_layout with
-        | Struct_field _ when not x.field_has_definition_issues ->
-            any "  \n" ++ const pp_size x.field_size
-        | _ ->
-            nop)
-    ++ any "  \n"
-    ++ const (option (any "Redefines:\n" ++ pp_cobol_block Cobol_ptree.pp_qualname')) x.field_redefines)
-  ppf x
+  let definition_has_issues = x.field_has_definition_issues in
+  let pp_qualname_opt_in_block' =
+    pp_cobol_block Fmt.(option ~none:(any "FILLER") Cobol_ptree.pp_qualname')
+  and pp_size ppf x =
+    match x.field_layout with
+    | Struct_field _ when not x.field_has_definition_issues ->
+        Fmt.fmt "  \n%a" ppf pp_size x.field_size
+    | _ ->
+        ()
+  in
+  match x.field_layout with
+  | Elementary_field _ when definition_has_issues ->
+      Fmt.(const pp_qualname_opt_in_block' x.field_qualname ++ any "\n\n" ++
+           any "*(layout omitted due to issues in item definition)*  \n" ++
+           const (option @@
+                  any "Redefines:\n" ++ pp_cobol_block Cobol_ptree.pp_qualname')
+             x.field_redefines)
+        ppf x
+  | _ ->
+      Fmt.(const pp_qualname_opt_in_block' x.field_qualname ++ any "\n\n" ++
+           const pp_field_layout x.field_layout ++
+           const pp_size x ++ any "  \n" ++
+           const (option @@
+                  any "Redefines:\n" ++ pp_cobol_block Cobol_ptree.pp_qualname')
+             x.field_redefines)
+        ppf x
 
 and pp_field_definition': field_definition with_loc Pretty.printer = fun ppf ->
   Cobol_ptree.pp_with_loc pp_field_definition ppf
@@ -235,18 +248,19 @@ let pp_renamed_item_layout: renamed_item_layout Pretty.printer = fun ppf x ->
       Fmt.const pp_struct subfields ppf x
 
 let pp_record_renaming: record_renaming Pretty.printer = fun ppf r ->
-  Fmt.(
-    pp_cobol_block (
+  let open Fmt in begin
+    pp_cobol_block begin
       const Cobol_ptree.pp_qualname' r.renaming_name
       ++ any "\nRENAMES "
       ++ const Cobol_ptree.pp_qualname' r.renaming_from
-      ++ const (option (any "\nTHRU " ++ Cobol_ptree.pp_qualname')) r.renaming_thru)
+      ++ const (option (any "\nTHRU " ++ Cobol_ptree.pp_qualname')) r.renaming_thru
+    end
     ++ any "\n\n"
     ++ const pp_renamed_item_layout r.renaming_layout
     ++ (match r.renaming_layout with
-    | Renamed_struct _ -> any "  \n" ++ const pp_size r.renaming_size
-    | _ -> nop) )
-  ppf r
+        | Renamed_struct _ -> any "  \n" ++ const pp_size r.renaming_size
+        | _ -> nop)
+  end ppf r
 
 let pp_record_renaming': record_renaming with_loc Pretty.printer = fun ppf ->
   Cobol_ptree.pp_with_loc pp_record_renaming ppf